fix: rename QVAULT files to lowercase
Renamed QVAULT.cpp and QVAULT.h to qvault.cpp and qvault.h respectively.
Updated the include statements in connection.cpp to reflect these changes.
This ensures consistency in file naming conventions.
