+++ qvault.cpp
